Add permissions to app.

+++ backend/src/resolvers/Mutation.js
@@ -2,14 +2,25 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: Check if they're logged in
+    // Check if they're logged in
+    if (!ctx.request.userId) {
+      throw new Error(`You must be logged in to do that!`);
+    }
 
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          // creates a relationship between the item and a user
+          user: {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
           ...args
         }
       },
@@ -42,8 +53,15 @@ const Mutations = {
     const where = { id: args.id };
 
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
     // 2. check if they own that item or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
 
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
@@ -113,7 +131,7 @@ const Mutations = {
 
   async requestReset(parent, args, ctx, info) {
     // 1. Check if this is a real user
-    const user = await ctx.db.query.user({ where: {email: args.email} });
+    const user = await ctx.db.query.user({ where: { email: args.email } });
 
     if (!user) {
       throw new Error(`No such user found for email "${args.email}"`);
@@ -130,6 +148,17 @@ const Mutations = {
 
     // console.log(res);
     // 3. Email them the reset token
+    const mailResponse = await transport.sendMail({
+      from: 'advancedReact@react.com',
+      to: user.email,
+      subject: 'Your Password Reset Token',
+      html: makeANiceEmail(`Your password reset token is here! 
+        \n\n <a href="${
+          process.env.FRONTEND_URL
+        }/reset?resetToken=${resetToken}">Click here to reset</a>`)
+    });
+
+    // 4. Return the message
     return { message: 'Thanks!' };
   },
 
@@ -144,7 +173,7 @@ const Mutations = {
     const [user] = await ctx.db.query.users({
       where: {
         resetToken: args.resetToken,
-        resetTokenExpiry_gte: Date.now() - 3600000,
+        resetTokenExpiry_gte: Date.now() - 3600000
       }
     });
 
@@ -161,7 +190,7 @@ const Mutations = {
       data: {
         password,
         resetToken: null,
-        resetTokenExpiry: null,
+        resetTokenExpiry: null
       }
     });
 
@@ -176,6 +205,38 @@ const Mutations = {
 
     // 8. Return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
